+++ src/main/java/me/drex/vanish/util/VanishedEntity.java
@@ -1,0 +1,9 @@
+package me.drex.vanish.util;
+
+public interface VanishedEntity {
+
+    boolean vanish$isVanished();
+
+    void vanish$setDirty();
+
+}